fix[all]: Sửa lại author bằng prePostAuthorize

+++ backend/src/main/java/com/react_online_judge/backend/service/ContestService.java
@@ -1,7 +1,9 @@
 package com.react_online_judge.backend.service;
 
 import com.nimbusds.jose.JOSEException;
+import com.react_online_judge.backend.configuration.SecurityConfig;
 import com.react_online_judge.backend.dto.request.ContestCreationRequest;
+import com.react_online_judge.backend.dto.request.ContestJoinRequest;
 import com.react_online_judge.backend.dto.request.ContestUpdateRequest;
 import com.react_online_judge.backend.dto.response.ContestResponse;
 import com.react_online_judge.backend.entity.Announcement;
@@ -22,6 +24,8 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.text.ParseException;
@@ -49,19 +53,22 @@ public class ContestService {
     AuthenticateService authenticateService;
     @Autowired
     private ContestPartipatorRepository contestPartipatorRepository;
-
+    @PreAuthorize("permitAll()")
     public ContestResponse getContestById(Long id) {
         Contest contest = contestRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.CONTEST_NOT_EXISTED));
         return contestMapper.toContestResponse(contest);
     }
+    @PreAuthorize("permitAll()")
     public ContestResponse getContestByTitle(String title) {
         Contest contest = contestRepository.findByTitle(title).orElseThrow(() -> new AppException(ErrorCode.CONTEST_NOT_EXISTED));
         return contestMapper.toContestResponse(contest);
     }
+    @PreAuthorize("permitAll()")
     public List<ContestResponse> getAllContests() {
         List<Contest> contests = contestRepository.findAll();
         return contestMapper.toContestResponseList(contests);
     }
+    @PreAuthorize("hasRole('ADMIN') and hasAuthority('CREATE_CONTEST')")
     public ContestResponse createContest(String token, ContestCreationRequest request) {
         Contest contest = contestMapper.toContest(request);
         User user;
@@ -84,6 +91,7 @@ public class ContestService {
             throw new AppException(ErrorCode.CONTEST_EXISTED);
         }
     }
+    @PreAuthorize("hasRole('ADMIN') and hasAuthority('UPDATE_CONTEST')")
     public ContestResponse updateContest(String token, Long id, ContestUpdateRequest request) {
         Contest contest = contestRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.CONTEST_NOT_EXISTED));
         User user;
@@ -103,14 +111,27 @@ public class ContestService {
             throw new AppException(ErrorCode.CONTEST_EXISTED);
         }
     }
+    @PreAuthorize("hasRole('ADMIN')")
     public List<ContestResponse> getMyCreatedContests(String token, Long userId) {
         List<Contest> contests = contestRepository.findByCreatorId(userId);
         return contestMapper.toContestResponseList(contests);
     }
+    @PreAuthorize("hasRole('ADMIN') and hasAuthority('DELETE_CONTEST')")
     public void deleteContest(Long id) {
-        contestRepository.deleteById(id);
+        Contest contest = contestRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.CONTEST_NOT_EXISTED));
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        try {
+            User user = authenticateService.getUserFromToken(username);
+            if (contest.getCreatorId() != user.getId()) {
+                throw new AppException(ErrorCode.UNAUTHORIZED);
+            }
+            contestRepository.deleteById(id);
+        } catch (ParseException | JOSEException e) {
+            throw new RuntimeException(e);
+        }
     }
-    public ContestResponse joinContest(Long contestId, String token) {
+    @PreAuthorize("hasRole('USER')")
+    public ContestResponse joinContest(Long contestId, String token, ContestJoinRequest request) {
         try {
             User user = authenticateService.getUserFromToken(token);
             Contest contest = contestRepository.findById(contestId)
@@ -118,7 +139,9 @@ public class ContestService {
             log.info("Participators: {}", contest.getContestParticipators().stream()
                     .map(p -> "User ID: " + p.getUser().getId())
                     .collect(Collectors.joining(", ")));
-
+            if (!contest.getPassword().equals(request.getPassword())) {
+                throw new AppException(ErrorCode.INVALID_PASSWORD);
+            }
             if (contest.getContestParticipators().stream()
                     .anyMatch(p -> p.getUser().getId() == (user.getId()))) {
                 throw new AppException(ErrorCode.ALREADY_JOINED);
@@ -137,5 +160,4 @@ public class ContestService {
             throw new AppException(ErrorCode.UNAUTHENTICATED);
         }
     }
-
 }
